refactor: change structure code to clean code (#532)
