+++ internal/common/exec/exec.go
@@ -1,4 +1,4 @@
-package common
+package exec
 
 import (
 	"encoding/json"
