+++ internal/common/log/log.go
@@ -1,8 +1,8 @@
 package log
 
 import (
-	"github.com/luismayta/envsecrets/v1/internal/app/config"
 	"github.com/luismayta/envsecrets/v1/internal/common/log/provider"
+	"github.com/luismayta/envsecrets/v1/internal/config"
 	"github.com/luismayta/envsecrets/v1/internal/errors"
 )
