+++ main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"github.com/luismayta/envsecrets/v1/cmd"
+	"github.com/luismayta/envsecrets/v1/cmd/envsecrets"
 	"github.com/luismayta/envsecrets/v1/internal/errors"
 )
 
 func main() {
-	err := cmd.Execute()
+	err := envsecrets.Execute()
 	if err != nil {
 		errors.Must(err, errors.ErrorUnknown, "Error in Execute")
 	}
